Moved observeLifecycleEvent from directory.
No need for directory for single item.
