+++ app/src/main/java/com/koleff/kare_android/ui/compose/screen/SettingsScreen.kt
@@ -18,7 +18,7 @@ import com.koleff.kare_android.common.navigation.NavigationEvent
 import com.koleff.kare_android.ui.compose.components.navigation_components.scaffolds.MainScreenScaffold
 import com.koleff.kare_android.ui.compose.components.SettingsList
 import com.koleff.kare_android.ui.compose.dialogs.EnableNotificationsDialog
-import com.koleff.kare_android.ui.compose.lifecycle_event.observeLifecycleEvent
+import com.koleff.kare_android.ui.compose.observeLifecycleEvent
 import com.koleff.kare_android.ui.view_model.BaseViewModel
 
 typealias SettingsViewModel = BaseViewModel
